+++ src/poa.rs
@@ -148,7 +148,7 @@ use petgraph::visit::Dfs;
 
 use petgraph::{Directed, Graph, Incoming, Outgoing};
 
-pub type POAGraph = Graph<u32, i32, Directed, usize>;
+pub type POAGraph = Graph<u64, i32, Directed, usize>;
 
 // Unlike with a total order we may have arbitrary successors in the
 // traceback matrix. I have not yet figured out what the best level of
@@ -158,7 +158,7 @@ pub const MIN_SCORE: i32 = -858_993_459;
 
 /// Trait required to instantiate a Scoring instance
 pub trait MatchFunc {
-    fn score(&mut self, a: u32, b: u32) -> i32;
+    fn score(&mut self, a: u64, b: u64) -> i32;
 }
 
 /// A concrete data structure which implements trait MatchFunc with constant
@@ -188,7 +188,7 @@ impl MatchParams {
 
 impl MatchFunc for MatchParams {
     #[inline]
-    fn score(&mut self, a: u32, b: u32) -> i32 {
+    fn score(&mut self, a: u64, b: u64) -> i32 {
         if a == b {
             self.match_score
         } else {
@@ -201,9 +201,9 @@ impl MatchFunc for MatchParams {
 /// can be instantiated using closures and custom user defined functions
 impl<F> MatchFunc for F
 where
-    F: FnMut(u32, u32) -> i32,
+    F: FnMut(u64, u64) -> i32,
 {
-    fn score(&mut self, a: u32, b: u32) -> i32 {
+    fn score(&mut self, a: u64, b: u64) -> i32 {
         (self)(a, b)
     }
 }
@@ -522,7 +522,7 @@ impl Traceback {
         &self.matrix[i][j]
     }
 
-    pub fn print(&self, g: &Graph<u32, i32, Directed, usize>, query: Vec<u32>) {
+    pub fn print(&self, g: &Graph<u64, i32, Directed, usize>, query: Vec<u64>) {
         let (m, n) = (g.node_count(), query.len());
         print!(".\t");
         for base in query.iter().take(n) {
@@ -588,13 +588,13 @@ impl Traceback {
 /// Uses consuming builder pattern for constructing partial order alignments with method chaining
 pub struct Aligner<F: MatchFunc> {
     pub traceback: Traceback,
-    query: Vec<u32>,
+    query: Vec<u64>,
     pub poa: Poa<F>,
 }
 
 impl<F: MatchFunc> Aligner<F> {
     /// Create new instance.
-    pub fn new(scoring: Scoring<F>, reference: Vec<u32>) -> Self {
+    pub fn new(scoring: Scoring<F>, reference: Vec<u64>) -> Self {
         Aligner {
             traceback: Traceback::new(),
             query: reference.to_vec(),
@@ -615,7 +615,7 @@ impl<F: MatchFunc> Aligner<F> {
     }
 
     /// Globally align a given query against the graph.
-    pub fn global(&mut self, query: Vec<u32>) -> &mut Self {
+    pub fn global(&mut self, query: Vec<u64>) -> &mut Self {
         self.query = query.to_vec();
         self.traceback = self.poa.global(query);
         self
@@ -635,9 +635,9 @@ impl<F: MatchFunc> Aligner<F> {
 pub struct Poa<F: MatchFunc> {
     scoring: Scoring<F>,
     pub graph: POAGraph,
-    pub node_index : HashMap<NodeIndex<usize>, u32>,
-    pub edge_index : HashMap<usize, (u32, u32)>,
-    pub seq_to_edges : HashMap<Vec<u32>, Vec<EdgeIndex<usize>>>,
+    pub node_index : HashMap<NodeIndex<usize>, u64>,
+    pub edge_index : HashMap<usize, (u64, u64)>,
+    pub seq_to_edges : HashMap<Vec<u64>, Vec<EdgeIndex<usize>>>,
 }
 
 impl<F: MatchFunc> Poa<F> {
@@ -649,9 +649,9 @@ impl<F: MatchFunc> Poa<F> {
     /// * `poa` - the partially ordered reference alignment
     ///
     pub fn new(scoring: Scoring<F>, graph: POAGraph) -> Self {
-        let mut node_index : HashMap<NodeIndex<usize>, u32> = HashMap::new();
-        let mut edge_index : HashMap<usize, (u32, u32)> = HashMap::new();
-        let mut seq_to_edges : HashMap<Vec<u32>, Vec<EdgeIndex<usize>>> = HashMap::new();
+        let mut node_index : HashMap<NodeIndex<usize>, u64> = HashMap::new();
+        let mut edge_index : HashMap<usize, (u64, u64)> = HashMap::new();
+        let mut seq_to_edges : HashMap<Vec<u64>, Vec<EdgeIndex<usize>>> = HashMap::new();
 
 
         Poa { scoring, graph, node_index, edge_index, seq_to_edges }
@@ -664,12 +664,12 @@ impl<F: MatchFunc> Poa<F> {
     /// * `scoring` - the score struct
     /// * `reference` - a reference TextSlice to populate the initial reference graph
     ///
-    pub fn from_string(scoring: Scoring<F>, seq: Vec<u32>) -> Self {
-        let mut node_index : HashMap<NodeIndex<usize>, u32> = HashMap::new();
-        let mut edge_index : HashMap<usize, (u32, u32)> = HashMap::new();
-        let mut seq_to_edges : HashMap<Vec<u32>, Vec<EdgeIndex<usize>>> = HashMap::new();
+    pub fn from_string(scoring: Scoring<F>, seq: Vec<u64>) -> Self {
+        let mut node_index : HashMap<NodeIndex<usize>, u64> = HashMap::new();
+        let mut edge_index : HashMap<usize, (u64, u64)> = HashMap::new();
+        let mut seq_to_edges : HashMap<Vec<u64>, Vec<EdgeIndex<usize>>> = HashMap::new();
         let mut entry = seq_to_edges.entry(seq.to_vec()).or_insert(Vec::<EdgeIndex<usize>>::new());
-        let mut graph: Graph<u32, i32, Directed, usize> =
+        let mut graph: Graph<u64, i32, Directed, usize> =
             Graph::with_capacity(seq.len(), seq.len() - 1);
         let mut prev: NodeIndex<usize> = graph.add_node(seq[0]);
         node_index.insert(prev, seq[0]);
@@ -693,7 +693,7 @@ impl<F: MatchFunc> Poa<F> {
     /// * `query` - the query TextSlice to align against the internal graph member
     ///
 
-    pub fn global(&mut self, query: Vec<u32>) -> Traceback {
+    pub fn global(&mut self, query: Vec<u64>) -> Traceback {
         assert!(self.graph.node_count() != 0);
 
         // dimensions of the traceback matrix
@@ -803,11 +803,11 @@ impl<F: MatchFunc> Poa<F> {
     /// * `aln` - The alignment of the new sequence to the graph
     /// * `seq` - The sequence being incorporated
     ///
-    pub fn consensus(&mut self) -> Vec<u32> {
+    pub fn consensus(&mut self) -> Vec<u64> {
         // If we've added no new nodes or edges since the last call, sort first
         let mut cns = Vec::new();
         for node in self.consensus_path().to_vec() {
-            cns.push(*self.graph.node_weight(node).unwrap() as u32);
+            cns.push(*self.graph.node_weight(node).unwrap() as u64);
         }
 
         cns.to_vec()
@@ -860,7 +860,7 @@ impl<F: MatchFunc> Poa<F> {
 
         cns_path
     }
-    pub fn add_alignment(&mut self, aln: &Alignment, seq: Vec<u32>) {
+    pub fn add_alignment(&mut self, aln: &Alignment, seq: Vec<u64>) {
         let mut prev: NodeIndex<usize> = NodeIndex::new(0);
         let mut entry = self.seq_to_edges.entry(seq.to_vec()).or_insert(Vec::<EdgeIndex<usize>>::new());
         let mut i: usize = 0;
@@ -930,7 +930,7 @@ mod tests {
     fn test_init_graph() {
         // sanity check for String -> Graph
 
-        let scoring = Scoring::new(-1, 0, |a: u32, b: u32| if a == b { 1i32 } else { -1i32 });
+        let scoring = Scoring::new(-1, 0, |a: u64, b: u64| if a == b { 1i32 } else { -1i32 });
         let poa = Poa::from_string(scoring, b"123456789");
         assert!(poa.graph.is_directed());
         assert_eq!(poa.graph.node_count(), 9);
@@ -939,7 +939,7 @@ mod tests {
 
     #[test]
     fn test_alignment() {
-        let scoring = Scoring::new(-1, 0, |a: u32, b: u32| if a == b { 1i32 } else { -1i32 });
+        let scoring = Scoring::new(-1, 0, |a: u64, b: u64| if a == b { 1i32 } else { -1i32 });
         // examples from the POA paper
         //let _seq1 = b"PKMIVRPQKNETV";
         //let _seq2 = b"THKMLVRNETIM";
@@ -956,7 +956,7 @@ mod tests {
 
     #[test]
     fn test_branched_alignment() {
-        let scoring = Scoring::new(-1, 0, |a: u32, b: u32| if a == b { 1i32 } else { -1i32 });
+        let scoring = Scoring::new(-1, 0, |a: u64, b: u64| if a == b { 1i32 } else { -1i32 });
         let seq1 = b"TTTTT";
         let seq2 = b"TTATT";
         let mut poa = Poa::from_string(scoring, seq1);
@@ -973,7 +973,7 @@ mod tests {
 
     #[test]
     fn test_alt_branched_alignment() {
-        let scoring = Scoring::new(-1, 0, |a: u32, b: u32| if a == b { 1i32 } else { -1i32 });
+        let scoring = Scoring::new(-1, 0, |a: u64, b: u64| if a == b { 1i32 } else { -1i32 });
 
         let seq1 = b"TTCCTTAA";
         let seq2 = b"TTTTGGAA";
@@ -998,7 +998,7 @@ mod tests {
 
     #[test]
     fn test_insertion_on_branch() {
-        let scoring = Scoring::new(-1, 0, |a: u32, b: u32| if a == b { 1i32 } else { -1i32 });
+        let scoring = Scoring::new(-1, 0, |a: u64, b: u64| if a == b { 1i32 } else { -1i32 });
 
         let seq1 = b"TTCCGGTTTAA";
         let seq2 = b"TTGGTATGGGAA";
@@ -1023,7 +1023,7 @@ mod tests {
 
     #[test]
     fn test_poa_method_chaining() {
-        let scoring = Scoring::new(-1, 0, |a: u32, b: u32| if a == b { 1i32 } else { -1i32 });
+        let scoring = Scoring::new(-1, 0, |a: u64, b: u64| if a == b { 1i32 } else { -1i32 });
         let mut aligner = Aligner::new(scoring, b"TTCCGGTTTAA");
         aligner
             .global(b"TTGGTATGGGAA")
